lib/utils/build.js now requires postBuild function in gatsby-node (#273)
* lib/utils/build.js now requires postBuild function in gatsby-node

Removes requirement for a post-build.js file in the root of the project for a custom post build step in gatsby build. Instead it requires for a postBuild export in the gatsby-node file in the root of the domain, where the webpack custom configuration is also stored.

* Add error for post build and update README

Added a better error message for custom post-build actions defined in gatsby-node.js, explains to user that file cannot be used.

Updated README.md by removing post-build.js option in the project structure and including a new section explaining how to include a post-build function for users.

* Ammend error message in post-build and README

Correct some errors in the README.

* Fixed linting error in lib/utils/build.js

Changed strings from double to single quotes to avoid eslint error.