Standardizing headers, tweaked wording
While re-reading this, I'm thinking the order of sections could be a bit hard to digest. (components, global styles, typography.js, plugins. @amberleyromo what order is sensible to you? Usually the most basic pathway is the best, so is there anything else we ought to cut out or re-order to make it more basic?