Typo: Return to wrong file
I was doing the gatsby.js tutorial and I was running into one issue, the tutorial said to return to index.js, but it should be to my-files.js, I eventually figured out that this was probably just a typo, but for a long time I kept changing my index.js file and didn't understand what I was doing wrong, turns out I was editing the wrong file. For so far I have enjoyed gatsby.js, I want to use it for my personal portfolio.