Automatically Create MDX Nodes from mediatypes
This changes the handling of user-created MDX nodes to unify the type
creation (all derived nodes are now of type Mdx) and also pulls the
creation into core. The transformation of any field with a compatible
mediaType is automatic now, requiring no user configuration. The
applicable mediaTypes are text/markdown and text/x-markdown, but are
configurable. MDX does not have it's own media type so files are still
handled by extension and the media type changes really only apply to
remote CMSs