chore: standardize on lowercase/dasherized file names (#10691)
## Description

I've seen some tweets that indicate that _sometimes_ there can be confusion as to whether to use uppercase or lowercase conventions for React component file names.

This isn't specific to Gatsby, but we can resolve some of this by documenting a consistent format, and for that I'd recommend that we consistently document lowercase with a dash separating terms (dasherized, e.g. NavBar is `nav-bar.js`). This is what we use in the starters (e.g. src/components/header.js instead of src/components/Header.js) outside of the blog starter--which has been normalized in this PR.

I've also done a simple find and replace for usage of these (e.g. src/components/Bio.js -> src/components/bio.js) so that the documentation matches the usage. There are additional uppercase components (e.g. src/components/SEO.js) but I left those as is, since the usage of them where referenced is also uppercase.

## Related Issues

Sorta related to gatsbyjs/gatsby#10671 as well as my comment [here](https://github.com/gatsbyjs/gatsby/pull/10690#discussion_r244221748)