fix(themes): Remove require.resolve in component resolve path (#11848)
This fixes [a bug](https://github.com/ChristopherBiscardi/gatsby-theme-examples/issues/21) where `.jsx` files could not shadow `.js` files. It
does this by removing the `require.resolve` call, which is node's
resolution algorithm that only works on .js files, with a continuation
of webpack's resolving hooks. This allows us to replace a .js file in
a theme with a .jsx, .tsx, .re, etc file in a site.

This branch can be tested in the gatsby-theme-examples repo using the [`jsx-extension-bug`](https://github.com/ChristopherBiscardi/gatsby-theme-examples/tree/jsx-extension-bug) branch