fix(gatsby-source-filesystem) Fix basic auth option on createRemoteFileNode. (#12507)
## Description

This PR fixes support for basic auth on `createRemoteFileNode`.   See Issue #12219 for a full description.  This is a regression that first appears in `gatsby-source-filesystem@2.0.23`

The change is isolated to `createRemoteFileNode`.  We need to handle the HTTP headers argument distinctly from the http client options.  I've added a new argument for passing the http client options.

## Related Issues

Fixes #12219