fix(gatsby): fix not focusing router wrapper after navigation (#13197)
* add probably failing test

* rename custom command to make it clear it's assertion

* first find page, then use router + import from @reach/router/es

* revert grabbing @reach/router/es directly

* use webpack alias to use @reach/router es build for browser bundles

* update snapshots to cover added id to focus wrapper

* testing

* mind blown

* use BASE_PATH and PATH_PREFIX

* fix focusing on navigation from 404 to different 404, improve focus managment tests

* fix most of focus navigation focus managment in development

navigating from and to 404 doesn't focus router wrapper correctly. This is because 404 in development is special case (see notes before skipped e2e-tests/development-runtime accessibility tests). This can be fixed but require a lot of changes which would be out of scope of this PR and would potentially make it harder to merge page manifest changes, so for now lets ship fixes that are already here.

* remove stray console.log

* add note about partial hydration when missing resources and hydrating

* update dev runtime

* update snapshot

* use nested route

* add matchPath to pageResources object

* update snapshot

* remove stray console.log