perf(gatsby-plugin-contentful): drop the super expensive fixids cycle (#27318)
perf(gatsby-plugin-contentful): drop the super expensive fixids cycle

BREAKING CHANGE: 
If you were relying on the `contentful_id` on the `sys` object in your queries  those are no longer exposed and you can safely change them to the `id` property on that `sys` object. 

If you were relying on the `id` property in the `sys` object, you should be aware that it is no longer "normalized". In particular, it will no longer get a 'c' prefixed to the id.